Reduce use of undeclared properties, use trait to track entities
A lot of the undeclared properties on this form & similar are just tracking
values of various relevant entities. This adds some functionality for that tracking
onto a trait & uses that. This reduces the need
to pass around values whose only importance is that they have
been loaded from the database & would require
another db call to access them from elsewhere.

Note I have swapped out the interaction with _contributorEmail & _contributorDisplayName
in getStatusMessage and send email receipts - these are the last 2 functions
called in the submit routine & the only reason the receipt function needs to set it is
to pass them to getStatusMessage. There are a couple of earlier interactions which are
odd so I ignored them for now.

The loop for sending out receipts is pretty messed up - so the first step is
getting the inputs & outputs for that loop clarified - but this
property is bang in the middle