Extract getDiscountID()
This is part of a general cleanup to stop passing the form into functions & then checking the form
for a possible property. Instead each form that calls the function has a consistent
(publicly supported) getDiscountID() function that it calls & discountID is passed in.

Note I was actually trying to give getPriceSetID this treatment but discountID was in the
way. In general I'm trying to make some consisent public functions available on forms
for retrieving relevant entity IDs rather than the current property / form
variable patchwork