Delete probably-unused TrackingFields.tpl
Was added for this issue: https://issues.civicrm.org/jira/browse/CRM-4493
With this rationale: http://civicrm.org/node/562
In this svn commit: https://github.com/civicrm/civicrm-svn/commit/4ea717

The proposal was to add a "CiviTracker" module that would take advantage of this template.
However, 14 years later the "CiviTracker" extension was never developed, and if it were made today
it would not need this template.