CiviCampaign - Move serialized contents of civicrm_survey.recontact_interval into civicrm_option_value.filter
Refactors out a thoroughly strange data structure.

Each option value associated with a survey has an associated recontact_interval, which is an integer
storing the number of days before the respondent should be recontacted.

Before: A serialized array was being stored in the survey with keys matching  the option LABEL!

After: There was an unused integer column, civicrm_option_value.filter which works fine for storing
the number directly for each option_value with no messy mapping needed.