Pass specific ids rather than ids array
It is unclear when we pass ids what ids we care about (& in general how many ids we are using from the
various functions). In fact 2 specific IDs are used (one from the _relatedObjects array) so pass
these in and stop passing in ids. Also stop setting an id that we never use