Declare CRM_Core_Exception in cleanDir
The only information our exception imparts is that it is
coming from CiviCRM Core  - but we should at least be consistent in that.

Also - I think passing something other than a string here would be bad - so
adding a type hint is good. We were already checking not target so the empty string check is redundant