Remove always true if
This removes an if that is always true

role_id is a required field on the participant form and the register from
search that extends it - so we don't need to check if it is filled out.

In addition we know that it is always an array because that's just the type of field it is.

Removing handling for pre-select 2 makes the code clearer...