Update offline event receipt (html version) to be more previewable & more consistent
This fixes the template work in message admin preview and also to show multiple
participants more consistently

Multiple participants are only relevant to this receipt in the edge case
scenario of sending a receipt to a participant who was registered front office
from the back office form. This was pretty wonky. As there is no other info on the form
for multiple participants I just did a block of rows for multiple participants