dev/core#4567 Make civicrm_mailing_event_queue.job_id nullable for verp purposes
This is a step towards some of the changes proposed in
https://lab.civicrm.org/dev/core/-/issues/4567

The basic ideal is that if civicrm_mailing_event_queue had the additional field/s
mailing_id
is_test
The it would be possible to prune data from civicrm_mailing_job (after
some field updating & query re-writing).

However, the first step along this path is to make it such that replies could
still be handled if the job_id had been nulled out.

This does that by ignoring job_id when checking if the queue_id & hash match, thus
loading the record that is being responded to.

The queue id is included when creating the hash but to be honest the
created hash is no more secure for the inclusion of the email_id or the
job_id. I added the SITE_KEY in cos that seemed like the sort of thing we do.

Provided the hash & queue_id are not a guessable combo (which
I don't think they are) having the job_id only adds illusory
extra benefits IMHO