Move nasty chunk of code to where it is used
This code to get the customFields array is tough going. I haven't unravelled it, just moved it from being
figured out in one function & used in another such that it is figured out in the
chunk of code it is used in.

This will probably fix some php 8.2 tests because rather than all tests
hitting it only those that actually have custom fields + receipts do.

It's a bit of a fake win but it will reduce fails & make further cleanup easier