The Battle of the Synonyms was put on-hold.
With the recent update, it fires a hook, and the hook-data describes the
queue-status as "deferred".

But the setting which drives it is called "queue_suspended".

Which word is better?  On reflection...  neither!

In CiviMail, there's some existing functionality where you can "pause" a
mail-blast.  And it's basically the same thing: we keep the queue-data, but
we tell the workers to temporarily refrain from execution.  And that's fine.
"Pausing" sounds less scary...