Afform - Switch to user-oriented breadcrumbs, move admin link to hover button
Before: In full-page mode, breadcrumbs were orieted toward form-building,
with links to edit the form and browse FormBuilder.

After: In full-page mode, breadcrumbs are oriented toward general users,
with links to parent-pages.

Parent pages are, by default calculated based off the route. This adds a 'title' key
in function `afform_civicrm_alterMenu` for the benefit of`CRM_Core_Menu::buildBreadcrumb`.

Alternately, if the Afform has been placed in the navigation menu, then the breadcrumb
is calculated based on the parent menu item.