Fix dev/core#4566 by ensuring groupNames are fetched
The problem was that EditUnsubGroupCtrl needed CRM.crmMailing.groupNames to be fetched,
but wasn't doing anything about it. Instead it was passively waiting around for someone
else to call for them to be loaded. The only place doing this was ViewRecipCtrl, and not consistently.
It was working in core but not in Mosaico.

I solved this by moving the loader functions to a service so they could be called from both places.