Fix rc regression - balance amount missing
This is pretty hard to get to (outside the test). You need to configure
an event with a pending contribution and then partially pay it
and go back & send a receipt.

Unfortunately the
Paid amount & Balance amount
 got removed from the html version by accident
and the tests didn't pick it up because they were testing
the text version - which had not had many of the recent updates
applied (it was going to be done after but we have since decided to remove the
text version).

This updates the test to check the html version, puts back the removed text
and adds it to the preview screen