Fix Participant form test to correctly test html version
We were checking for text that is present on the text version - but not the html.

In this case the html version is right - ie we only show paid amount & balance if
relevant  - ie if part payment has been made. If it is all paid or all unpaid we don't