Follow up cleanup on multiple participant form edit
In testing on the recent regression some code turned out to
be unreachable.

In the scenario were we have a pending pay later registration for multipl
participants we find that the field priceSetId is empty when submitted
and the field total_amount must be submitted, if it is submitted.

Therefore we find that a) any value added to total_amount in the
previously-shared preparePaidEventProcessing is overwritten
shortly after by the submitted value
and hence should not be set in this flow in preparePaidEventProcessing.

Also the code that sets additionalDetails is only hit when priceSetID is
not empty but it never is per above. So that code is setting an unused variable