Move set up of complicated price set to the ContributionPageTestTrait & test to the confirmTest
We need to add some more testing but most of our existing tests are within the apiv3 test harness
which is not a great location. This moves one test over to the full form test flow
and also pimps the helper trait to support this complex price scenario (with fields
of a mixture of types & financial types.

I want to add some checks on the email output but this specific test is pay later
so I won't do it in this PR