dev/core#4740 - Fix missing tagsets when label doesn't match name
CRM_Core_Form_Tag::buildQuickForm was previously being passed an array of [id => name]
but it was wastefully ignoring the passed ids and looking it up from the name.

Now due to bcf668d is being passed an array of [id => label] which is all it needs,
just had to remove the wasteful id lookup when it's already there.