Fix notice regression on contact edit
This notice occurs when the contact has one or more subtypes allocated. Note that the thing
I focussed on while testing was ensuring that it was possible to add a contact sub-type
and save custom field data applicable to that sub-type. I also checked with the subtype
name differing from the label