Fully remove interaction with _params from participant form
This removes the last places _params was accessed & adds a __get for any external access to it

Note this also removes the expiry date from the template. I can't find
the discussion we had about this (maybe on chat) but it was generally agreed that
since we hadn't added a column for this in the DB it made sense to exclude from
the template (as the template should be the same however it is sent).
There was some difference of opinion about whether it should have
stored in the DB though