dev/core#2607 - Add deleted-files-list
Towards addressing the problem that some upgraders leave old files in the tree, which
can cause php fatal errors when they continue to autoload, or security issues if the
files were deleted for security reasons.

This generates a list of files which no longer exist. We could use that list to attempt
to auto-remove them during upgrades, or at the very least show an alert if they are present.