DAO - Remove fieldKeys(), getLog() and getTableName() from template
In 2020, 9d89b14 added these functions to the base class, effectively making them redundant in the generated DAO files.
But out of an abundance of caution, they were not immediatetly removed from the DAO template.
This was so that extensions that generate DAOs with civix (which uses the same template from core) would
not immediately lose backward-compat with prior versions of Civi.
At this point enough time has gone by. The functions were added to 5.31 which is quite a few ESR-releases old now.
Any extension wishing to generate DAO files compatible with versions < 5.31 can do so by using an older version of core
when running civix:generate.