dev/core#4674 - Add Oauth links for AdminUi MailSettings
This fixes https://lab.civicrm.org/dev/core/-/issues/4674 by adding a LinksProvider
for the new GetLinks action from #27973.

It uses the CRM_Core_BAO_MailSettings::getSetupActions() function to generate the extra links,
but there is a mismatch between the two functions:

- getSetupActions() returns a fully-formed url
- API::getLinks() returns a path

However, I noticed that while in theory hook_civicrm_mailSetupActions could return various urls,
in practice they are all the same: a generic civicrm/ajax/setupMailAccount redirect.
So this returns that as the path and it works.

Perhaps we should deprecate the option to return a url from hook_civicrm_mailSetupActions
(or just remove it, since it appears to be unused & YAGNI).