CRM_Utils_System::url() - Respect other default schemes
The `System::url()` entry-point has a contract designed for a "two scheme
world" (frontend or backend). We're entering a "multi-scheme" world
(frontend or backend or web-service or oembed).

This patch keeps the behavior of `url()` basically untouched.  However, if
you have setup some other scheme as the "default/current", then it will
reroute the request.

Before+After:

* `System::url()` with a specific flag (`$frontend` or `$forceBackend`) behaves the same.
* `System::url()` with no specific flag -- and in a default environment -- behaves the same.

Changed:

* `System::url()` with no specific flag -- and in an alternatee environment -- routes
  through that environment.