CRM_Utils_System::url() - Respect other default schemes
The `System::url()` API has a contract designed for a "two scheme world"
(frontend or backend).  The gist of the contract is this:

- By default, the URL is rendered in the *current* environment (frontend xor backend)
- You may optionally request a specific environment (`$frontend=TRUE` or `$forceBackend=TRUE`).

We're entering a "multi-scheme" world (with frontend or backend; and also:
web-service or oembed).  This is more apparent in the `Civi::url()` API,
where you can give a scheme (`current://` or `frontend://` or `oembed://`).
But we still have many calls to `System::url()`.

The patch preserves the behavior of `System::url()` for all normal work in the existing
frontend/backend environments. But it adds support for other "current" environments:

==> If you're processing an oembed page-view and call `System::url()` (with
    default options), then it should return the oembed-style URL.

Of course, if you give an explicit flag (`$frontend`, `$forceBackend`), then that
will still be respected.