SearchKit - Fix wrong column being removed from displays
Fixes recent regression from 4099cd6 causing the wrong column to be removed from a display.
To reproduce: Create a search display, go back to the main search criteria and delete a column,
then tab back to the display. The deleted column should have been removed but instead the wrong
one is still there.