Use distinct test workflow message for workflow message tests
The tests for the workflow message subsystem use an instance of a test
case_activity template attached the real message workflow template. They
test the text version - meaning that removing it, like we have with
others, causes various fails. In addition adding a real template that
does not match the test version causes mayhem.

In this case we want to ensure we keep all the tests for the subsystem
functionality but undo the tethering to a particular
template to allow that template to have updates without
having to go deep into this code