Standardise implementation of financial type acl in query object
This fixes it to call the selectWhere in the BAO_Query, as apiv4 does. There is test
cover in the touched test.

The primary goal here is to get this logic out of core & to disabled this extension
on most sites with a view to moving the whole extension out of core in time