dev/core#4907 Include participant role filter in apiv4 available spaces calculation
Per https://lab.civicrm.org/dev/core/-/issues/4907
there are a bunch of inconsistencies in the determination of
how many spaces remain on the event. This addresses 1 - ie
the apiv4 does not give due attention to whether the role should be counted or
not. It also does some refactoring in the eventFull function, towards future sanity