UFGroup - Cleanup getCustomFields and standardize checkPermission param
The caching in this function is redundant with the cache already in the function it calls, so I removed it.
The checkPermissions param sometimes comes in as an odd value based on profile permissions,
so this forces it to only be one of the two available options, defaulting to the stricter one.