Permission - Catch runtime exception if permission hook cannot be called
Previously CRM_Core_Permission::check did not rely on calling hook_civicrm_permission and could run at any time.
Now in order to parse implied permissions, we do call the hook and this causes problems with the upgrade tests.
In order to keep compatability with the status-quo, this catches the exception, which means implied permissions
cannot be evaluated at those odd times. OTOH, is there even a user or session yet during those odd times so does it matter?