DAO - Remove buggy caching from import/export
Back in 84a0493c Tim noticed that the existing function had a caching bug where the $prefix mode was
ignored by the cache so whatever mode was called first would be cached and returned to all
subsequent callers of the function.

Presumably to expedite an already huge refactor, he kept the bug, adding a commented-out fix
with the note for posterity: "Bug-level compatibility -- or sane behavior?"

Since the buggy behavior hasn't resulted in any reported problems in all these years,
I'm guessing that these functions aren't called very often (and never with opposite values
for $prefix during the same request).

I think the sane behavior is to just remove the cache entirely, as it achieves practically nothing:
all these functions do is filter a small array (returned from the already-cached fields() function),
and that's such a trivial operation in PHP it really doesn't need to be cached.