Afform - Fix incorrect html encoding when saving
Fixes  https://lab.civicrm.org/dev/core/-/issues/4977

This reverts 0ae1b99 AND 4a439e7 back to the original version which used a simple str_replace.

The problem with 4a439e7 is that function mode is deprecated in php 8.2.
The problem with 0ae1b99 is that it was wrong and broke the html encoding of the file.

So this takes us all the way back to the original version of the function, which worked fine for what it did.