Pass values into DedupeRule->sql() rather than hybrid object
It doesn't really make sense for this query calculation function to be on a different class
I kinda get what they were thinking but the way our BAOs are half a collection of static
functions but occassionally get serious about being an object but then
like extra properties set without explanation does us no favours. In this
case it would be most sane if all the query calculation was on the same class
(probably a new one but for now just looking at one of the existing ones). To
do that we need to get clearer about the input the sql() function needs - this switches
it most of the way to being static-ready & move-ready by passing in the values as arrays
rather than setting them as properties