Import - Stop using headerPattern from field metadata as it is not localized.
The 'headerPattern' key returned from DAO::fields() is only used in one place - import.
And while some of the patterns are quite clever, the only ever worked in English.

So this solution is not going to be quite as clever in English,
but at least it will work in other languages.

Note that these headerPatterns are only used as suggestions - the user still gets to
manually choose the field mapping (and until now, non-English users would always have to).

This will allow us to rip out the crufty headerPattern stuff from field metadata,
which aside from this one place is unused in `universe` (except CiviHR which is abandoned).