Fix js error when clearing the template on SMS trait
The check on document.getElementById('subject').length)
was causing a js error when not present - whereas the jquery version quietly does nothing if not present

dev/core#5029 somewhat fix save template on sms trait

This at least ensures it is possible to update the template & also fixes the notice.

It also fixes a js error when clearing the template if the subject has not been filled in

It still seems a little funky but I think it is closer to right & the gap is not on not
doing enough hiding rather than not showing what is needed.

I also fixed the save to use the api & pulled in https://github.com/civicrm/civicrm-core/pull/29505
because I allowed the permissions to be FALSE on save so wanted this protection.

Note I am targetting the just-cut 5.72 rc as the SMS form has been substantially re-written in
5.72 so my goal was to resolve the other known issues in 5.72 rather than spread changes
to this form over more than one release