Afform - Fix boolean options
Fixes https://lab.civicrm.org/dev/core/-/issues/4753

The previous code was taking care to cast option values to true/false for
boolean fields, but this was then undone by the two-way binding on the field.
By breaking the binding we can update the values in a way that won't be
undone by the Angular bindings.