Fix dedupe query wrangling to combine queries where possible
This addresses a performance issue we have been seeing in finding duplicates
using a name + address rule. It used to perform 'ok-ish' but a while back it suddenly
flipped to become 'server-killing'. I turns out that previous attempts
to optimise dedupe queries used a very crude mechanism whereby the weight
that fields had was ignored in favour of running queries first on the tables with
less rows. In our case the performance flip happened when the number
of rows in civicrm_address passed civicrm_contact - ie went from looking for
duplicates on
- street_address
- first_name
- last_name

to
- first_name
- last_name
- street_address

Since subsequent queries are constrained by earlier queries this
created a situation where it fell apart.  Testing the first query on
a small-ish group of 5000 contacts I had to kill it after 58 minutes.

Looking at fixing it just flipping it back would not
have been consistently better but the situation was dramatically better if
I combined the queries ie

- just first_name = > 1 hour
- first_name + last_name in the same query = 144 seconds
- first_name + last_name in the same query + a combined index = 23 seconds
- first_name + last_name + street_address (no index change) in the same query = 1.8 seconds

I haven't nailed that last one in code yet but this does the first_name + last_name
in the same query.

THere are 2 parts to the fix

1) figuring out what queries can be combined
2) combining them

**Figuraing out what queries can be combined**
This is done using the concept of a [power set](https://www.oreilly.com/library/view/php-cookbook/1565926811/ch04s25.html)
to get all the possible combinations, which are then whittled down to the ones that meet the threshold.
There is some fairly solid testing on this in isolation in this PR.

**Combining them**
This is covered in this PR ONLY for fields in the same table. I'm gonna try to follow up
by figuring out the latter. It is also only done, at the moment, when the
parameters are NOT passed (which would be the case when checking for duplicates
before adding a contact).

I have used an opt-in define for this. I hope it will be temporary but
I feel like it would be good to get wider feedback before fully making the switch