Finish consolidating customData code on the Contact form
Note this will resolve the php8.2 issue as _groupTree is an internal variable rather than
an undefined property after this. It probably makes sense to follow up with this
and other forms that display custom data such that the custom data is loaded always
rather than just sometimes by ajax. It would make for simpler saner code