Fix help function to not call fetchWith
The fetchWith function is not part of the native smarty object - rather it is part of our CRM_Core_Smarty.

In Smarty5 the smarty object that is passed here is
an actual Smarty object - not a CRM_Core_Smarty object so we can't call functions that
extend CRM_Core_Smarty. This was also the case in GenCode and Setup although
we somehow got it to work it was really painful & not trying
to call functions that hack-extend Smarty within Smarty would have helped a lot
last time too....