TokenCompatSubscriber - Consolidate chained regex. Use more singular regex.
Chaining a series of independent regex's can make it a bit harder to think about - e.g. you have
consider the impact of sequencing.

Calling `preg_replace()` with an array means that the order won't affect the outcome, but it still
allows you write each regex as a distinct item.