Remove mailing_backend_store, confusingly set when we switch the environment to Development
When we set the environment from Production to Staging or Development, the SMTP settings
are copied from mailing_backend to mailing_backend_store. Then when a Scheduled Job is
run with runInNonProductionEnvironment=TRUE, the setting is copied from _store back to
mailing_backend, so that emails can be sent, if necessary. If the Job does not fatal,
the settings are later on reset to what they were.

However, there is no way to set the backup SMTP settings and the admin is not informed
about this. An admin may also want to execute a Scheduled Job and have the emails
redirected to DB.

Conclusion: I hesitated between adding a complex UI to have "dev" and "production"
SMTP settings, but I think that assuming that running a Job in runInNonProductionEnvironment
does not necessarily imply that I want to send emails as production. So I recommend we
remove this behaviour completely.

Co-Authored-By: DemeritCowboy <demeritcowboy@hotmail.com>