afform - fix - set default value for a select field by url
Overview
----------------------------------------
After updatinging to > ~5.65 it is not longer possible to set default values for a select field by url if a format like this is used: civicrm/forms/test#?event_type_id=1,2,3
Currently tested with 5.72.3.

Before
----------------------------------------
no default values set

After
----------------------------------------
working again

Technical Details
----------------------------------------
correctValueType returns for '1,2,3' alwayes NaN if we have the combination of isMultiple and data_type = Integer instead of an array of integers. This change moves the split befor correctValueType and it works again.