Remove the Participant Role from Event confirmations
For most forms and emails, the Participant Role is not something that the user
selects, and is usually not relevant to them. CiviCRM is being overly pedantic
with this information, which is probably not relevant to 90% of people, and I
don't think we should annoy them because 10% need it.

Also, it didn't support translation most of the time, because it uses the label
instead of the name (and I didn't feel it was worth fixing). In any case, people
might not use "Attendant" but rather some custom option, and then CiviCRM would
behave differently for some obscure reason.