Add mailSettingId parameter to emailProcessor hook
This adds a new mailSettingId parameter to the emailProcessor hook.

The emailProcessor hook is called after inbound emails are processed
and provides parameters related to the email and result/created entity.

However, there is currently no way to figure out which mail setting
(mail account)  the email relates to - the created entities themselves
don't include a FK to the mail setting, and the mail object itself
does not help either since it doesn't give us access to the (trusted)
mail envelope which would contain the canonical From/To header.