Standalone Layout - More evocative "upload(s)" folders.
For a moment, I really liked the symmetry of "public/uploads" and
"private/uploads".  And could be talked back into it.  It almost cleans up
the mess of "upload" terminology.

But I don't think it quite succeeds. The pre-existing mess is too much.

* The problem is "uploadDir": so there are two "upload(s)" folders, and
  neither is configured with "uploadDir".
* When you migrate from old-layout to new-layout, you'll have "upload(s)" in the
  before+after layouts... but the substance has to get swapped around.
* The programmatic default for `[civicrm.files]` and `[civicrm.private]`
  in the same place. If that happens, then the "uploads" would be simultaneously
  public+private. Obviously, we're discouraging that. But $wildInternet.

So I was looking for phrases that (a) would be more distinctive, (b) would have
some resonance. I think this does the job:

* "public/media" (images, videos, PDFs, etc that you're publishing;
  as in the CKEditor/KCFinder dialogs)
* "private/attachments" (files that you put onto a specific node;
  web UI calls these "Attachments" for activities and for civimail)