Add nofilter to templates for avoiding default escaping
This is the documented method for Smarty5 & not documented for SMarty2 but appears to work on both

https://smarty-php.github.io/smarty/stable/programmers/api-variables/variable-escape-html/

It is not possible to enable default escaping on Smarty 3 or 4 & there are no production uses
on Smarty2 so the goal is to make this viable for Smarty5.

Notably Smarty5 seems to require it when json_decode is invoked but Smarty2 did not.

This PR is incomplete but it seems that we could do a mass update where smarty::nodefaults
is used to use the nofilter syntax & if this approach is agreed I can follow up with that