afformAdmin - Ensure conditional values have correct data type
This was a big fight between me & AngularJS.
It had been incorrectly formatting field values & I eventually tracked the problem
to a load-order issue, where AngularJS provides its own formatter (`stringBasedInputType`),
which casts the value to a string and ruins the serialization of arrays. After much trial-and-error
I figured out that formatters are applied in reverse-order, and for some reason the angular core
formatter is added quite late. The only way I could find to add one even later was to use a postLink
function in the directive.