dev/core#5306 dev/core#5308 Participant.validate() pseudocode
I've been saying what we REALLY need to start on any form of Participant.validate action is
some type of flow chart to identify what validating a participant means.

Monish has been doing some great work on what a membership.validate api
might look like & I have been plagued by not knowing if the signature would
work for participant & absent any real discovery on what that flow is have felt at sea.

I don't do flow charts but I figured I could pseudocode out the flow, which is what this
PR is. It might form a starting point for where to go with this action but obviously the
tests would be the most important part of any eventual PR.

Note that one conclusion I reached is that the only context we really need here is
whether the intent is to do an administrative registration or not - which combined
with the permission will allow overriding of limits etc.

The other thing I *though* would be context is whether they are trying to finish
an approved or pending from waitling transaction but working through this I feel
the presence of the participant id will allow us to check if they have a status
that is eligible for the update