Follow up to #30415 - restrict permissions
In #30415 we removed a requirement to restrict permission to editing a
price set Access CiviEvent, leaving just access CiviCRM.

Since a price set will either be used on a Contribution page or
an event page it makes more sense to require the same permissions
as adding/editing an event or contribution page.

This tightens up the permissions using the ';' syntax documented
https://docs.civicrm.org/dev/en/latest/framework/routing/#access_arguments
which is in use elsewhere.