Backdrop - Fix MockPublicFormTest.  Ensure session is committed during direct.
Overview
--------

This fixes a failing test. The failure does indicate a buggy behavior.

The fix fills-in a missing method `CRM_Utils_System_Backdrop::onCiviExit()`.
The method is nearly identical to the one in `CRM_Utils_System_Drupal`.

Use-Case
--------

In MockPublicFormTest, we send email with an authenticated link to a form, eg

https://example.com/civicrm/my-form?_authx=...&_authxSes=1

During this request, it authenticates the token and stores the contact ID in the session.

Finally, it issues a redirect to reload `civicrm/my-form` with a clean URL.

Before
------

Within PHP, it stores the data in `CRM_Core_Session`... but that data NOT written persistently.

After
-----

It stores the data in `CRM_Core_Session`... and actually writes it persistent.