Remove unreachable code.
This code is hard to follow back but ultimately this function is only called from Flexmailer Abdicater::onCheckSendable if
sms_provider_id is not empty - so all the code that does not related to the sms_provider_id is
unreachable. This removes that code & makes the calling function clearer / easier
to trace.

This is legacy complexity being removed