Use usage to determine which contact fields can be imported for contribution import
This updates the getFields retrieval to filter on the usage being set to import
rather than on the field type. This is more correct and also supports
adding additional fields through extensions - ie the Deduper
extension adds 'full_name' as a field for 'create/update' and when
this is provided it uses a library to split it into parts it
a somewhat correct way. Being able to access this via import
is helpful for users who have data in this format.

Currently this function is only called via the Contribution.import
and most of the fields are filtered out if civiimport
is not enabled.

Note that there are some fields that are no longer available to
import via the contribution import with this patch
- addressee_id
- is_deleted
- email_greeting_id
- postal_greeting_id
- preferred_mail_format

These do not have the usage of 'import' in the schema
and I think they should have always been filtered out in this context.

I went back & forth on
- also filtering readonly = false (should be implicit in the usage IMHO)
- setting the action to 'save' - I think this makes sense as these
are importable fields & I don't want to have to make my create/update only
field also apply to get which it seems to default to (rather than
'action_unspecified'