Remove use of createAPIError (gets swallowed in php8.x
This is called with the modeException so it throws an exception - but with NULL for the
data - which should be an int.

It winds up in a deprecation loop & the error is lost

This is hard to hit but if you delete the file after it has been
identified it is hit. Since it's in CRM code one of our errors makes sense