dompdf - Abide by configurable path to "upload" dir
Overview
-------------

This changes the calculation for the "getCacheDir()".  On most systems, it
should behave the same.  However, on some systems (such as Standalone) the
"upload" dir has been renamed -- and the old formula pointed to the wrong
place.

With this revision, it abides by the local configuration.

Before
------

* Various unit-tests involving PDFs fail if they run on Standalone. Ex:
    * `api_v3_JobTest::testMailReportForPdf`
    * `CRM_Activity_Form_Task_PDFTest::testCreateDocumentUnknownTokens`
    * `CRM_Contribute_Form_Task_PDFLetterCommonTest::testGroupedThankYous`
* On Standalone, the default value of `getCacheDir` is invalid.

After
-----

* Those unit-tests pass on Standalone.
* On Standalone, the default value of `getCacheDir` is valid.

Comment
-------

The problem for "Standalone" is that "[civicrm.files]/upload" was renamed to
"[civicrm.private]/tmp", but `getCacheDir()` was hard-coded to the other
convention.  This formula should work with either.